requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via -r requirements.in
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
distro==1.9.0
    # via openai
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.5
    # via fastapi
google-auth==2.34.0
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via
    #   -r requirements.in
    #   uvicorn
httpx==0.27.2
    # via
    #   fastapi
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
openai==1.34.0
    # via -r requirements.in
orjson==3.10.7
    # via fastapi
packaging==24.1
    # via gunicorn
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.9.1
    # via
    #   fastapi
    #   openai
pydantic-core==2.23.3
    # via pydantic
pygments==2.18.0
    # via rich
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via -r requirements.in
rich==13.8.1
    # via typer
rsa==4.9
    # via google-auth
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.37.2
    # via fastapi
tinydb==4.8.0
    # via -r requirements.in
tqdm==4.66.5
    # via openai
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
    #   uvicorn
ujson==5.10.0
    # via fastapi
urllib3==2.2.3
    # via requests
uvicorn[standard]==0.30.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
uvloop==0.19.0
    # via
    #   -r requirements.in
    #   uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
